pyproject.toml
[tool.poetry]
name = "nade"
version = "0.2.7"
description = "Natural affect detection allows to infer basic emotions from social media messages"
authors = ["Christian Hotz-Behofsits <chris@hotz-behofsits.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://nade.rds.wu.ac.at"
repository = "https://github.com/inkrement/nade"

[tool.poetry.dependencies]
python = ">= 3.6"
pybind11 = ">= 2.0"
fasttext-wheel = ">= 0.9.2"
numpy = ">= 1.19"
lightgbm = ">= 3.2"
pyarrow = ">= 6.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/inkrement/nade/issues"
